Fatras: move photon conversion
